Rename models and repos modules
