+++ trac-runtime/python/src/trac/rt/exec/engine.py
@@ -20,7 +20,7 @@ from dataclasses import dataclass, field
 
 import trac.rt.config as config
 import trac.rt.impl.util as util
-import trac.rt.impl.model_loader as _models
+import trac.rt.impl.models as _models
 import trac.rt.impl.storage as _storage
 
 import trac.rt.exec.actors as actors
